Bunch of files purged to the glory of linter & AppRouter refactored
