+++ client/src/App.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import AppRouter from './components/AppRouter.js';
+import AppRouter from './components/AppRouter.jsx';
 import Header from './components/Header/Header';
 import Footer from './components/Footer/Footer';
-import NavBar from './components/NavBar/NavBar.js';
+import NavBar from './components/NavBar/NavBar';
 import './styles.css';
 
 function App() {
