+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './redux/store';
-import App from './App.js';
+import App from './App.jsx';
 
 ReactDOM.render(
   <Provider store={store}>
